// src/users/dto/create-user.dto.ts
export class CreateUserDto {
    id: string ;
    addr1: string;
    addr2: string;
    city: string;
    comment: string;
    country: string;
    email: string;
    image: string;
    name: string;
    password: string;
    phone: string;
    role: string;
    state: string;
    username: string;
    zip: string;
}